Fix Encrypt & Decrypt
