+++ practices/level1/p07_encrypt_decrypt/p07.h
@@ -1,6 +1,5 @@
-#ifndef _ENCRYPT_H
-
-#define _ENCRYPT_H
+#ifndef _P07_H
+#define _P07_H
 
 #include <stdio.h>
 #include <string.h>
@@ -30,4 +29,15 @@ int encrypt(char input[])
     return (int)key;
 }
 
-#endif // _ENCRYPT_H
+void decrypt(char input[],int key){
+    int len = strlen(input);
+    for(int i = 0; i < len; ++i){
+        if(input[i] <= 'z' && input[i] >= 'a'){
+            input[i] = 'A' + (input[i] - 'a' + CHARSET_SIZE - key) % CHARSET_SIZE;
+        }else{
+            input[i] = 'a' + (input[i] - 'A' + CHARSET_SIZE - key) % CHARSET_SIZE;
+        }
+    }
+}
+
+#endif
